+++ common/src/main/java/com/github/cao/awa/conium/bedrock/event/after/BedrockWorldAfterEvents.kt
@@ -2,7 +2,7 @@ package com.github.cao.awa.conium.bedrock.event.after
 
 import com.github.cao.awa.conium.annotation.bedrock.BedrockScriptApi
 import com.github.cao.awa.conium.annotation.bedrock.BedrockScriptApiFacade
-import com.github.cao.awa.conium.bedrock.event.after.item.use.BedrockItemUseOnAfterEvent
+import com.github.cao.awa.conium.bedrock.event.after.item.use.on.BedrockItemUseOnAfterEvent
 
 @BedrockScriptApi
 @BedrockScriptApiFacade("WorldAfterEvents")