+++ common/src/main/java/com/github/cao/awa/conium/bedrock/event/context/item/use/BedrockItemUseEventContext.kt
@@ -1,0 +1,22 @@
+package com.github.cao.awa.conium.bedrock.event.context.item.use
+
+import com.github.cao.awa.conium.annotation.bedrock.BedrockScriptApi
+import com.github.cao.awa.conium.annotation.bedrock.BedrockScriptApiFacade
+import com.github.cao.awa.conium.bedrock.entity.player.BedrockPlayer
+import com.github.cao.awa.conium.bedrock.event.context.BedrockEventContext
+import com.github.cao.awa.conium.bedrock.item.stack.BedrockItemStack
+import com.github.cao.awa.conium.bedrock.world.BedrockWorld
+import com.github.cao.awa.conium.bedrock.world.dimension.BedrockDimension
+
+@BedrockScriptApi
+@BedrockScriptApiFacade("ItemUseBeforeEvent", "ItemUseAfterEvent")
+class BedrockItemUseEventContext(
+    scriptSource: Any,
+    val world: BedrockWorld,
+    val itemStack: BedrockItemStack,
+    val source: BedrockPlayer
+) : BedrockEventContext(scriptSource) {
+    var cancel: Boolean = false
+
+    override fun world(): BedrockWorld = this.world
+}