+++ common/src/main/java/com/github/cao/awa/conium/bedrock/event/context/item/use/BedrockItemUseOnEventContext.kt
@@ -3,14 +3,14 @@ package com.github.cao.awa.conium.bedrock.event.context.item.use
 import com.github.cao.awa.conium.annotation.bedrock.BedrockScriptApi
 import com.github.cao.awa.conium.annotation.bedrock.BedrockScriptApiFacade
 import com.github.cao.awa.conium.bedrock.entity.player.BedrockPlayer
-import com.github.cao.awa.conium.bedrock.entity.player.toBedrock
+import com.github.cao.awa.conium.bedrock.entity.player.bedrockPlayer
 import com.github.cao.awa.conium.bedrock.event.context.BedrockEventContext
 import com.github.cao.awa.conium.bedrock.item.stack.BedrockItemStack
-import com.github.cao.awa.conium.bedrock.item.stack.toBedrock
+import com.github.cao.awa.conium.bedrock.item.stack.bedrockItemStack
 import com.github.cao.awa.conium.bedrock.world.BedrockWorld
 import com.github.cao.awa.conium.bedrock.world.dimension.BedrockDimension
-import com.github.cao.awa.conium.bedrock.world.dimension.bedrock
-import com.github.cao.awa.conium.bedrock.world.bedrock
+import com.github.cao.awa.conium.bedrock.world.dimension.bedrockDimension
+import com.github.cao.awa.conium.bedrock.world.bedrockWorld
 import net.minecraft.item.ItemUsageContext
 import net.minecraft.server.network.ServerPlayerEntity
 
@@ -28,13 +28,13 @@ class BedrockItemUseOnEventContext(
     override fun world(): BedrockWorld = this.world
 }
 
-fun ItemUsageContext.toBedrock(scriptSource: Any, source: ServerPlayerEntity): BedrockItemUseOnEventContext {
+fun ItemUsageContext.bedrockEventContext(scriptSource: Any, source: ServerPlayerEntity): BedrockItemUseOnEventContext {
     val context = BedrockItemUseOnEventContext(
         scriptSource,
-        this.world.bedrock,
-        this.world.server!!.bedrock,
-        this.stack.toBedrock(),
-        source.toBedrock()
+        this.world.bedrockDimension,
+        this.world.server!!.bedrockWorld,
+        this.stack.bedrockItemStack(),
+        source.bedrockPlayer()
     )
     BedrockEventContext.contexts[scriptSource] = context
     return context